Improve support for mobile devices
Smaller devices with touch screen doesn't trigger hovering.

This solution add a hover class to the business card if it's clicked.
It's not optimal - clicks outside the card should close it. Will work
more on improving this. Got a tip on how to detect touch screens [1].

Trying to avoid including javascript libraries like jQuery, so had a
little help with writing native javascript [2].

[1] http://www.prowebdesign.ro/how-to-deal-with-hover-on-touch-screen-devices/
[2] http://toddmotto.com/is-it-time-to-drop-jquery-essentials-to-learning-javascript-from-a-jquery-background/